Moved root route into root module
